mudando csv para classes
