+++ src/main/java/br/com/hbsis/linhaCategoria/LinhaCategoriaService.java
@@ -1,23 +1,10 @@
 package br.com.hbsis.linhaCategoria;
-
-
-import br.com.hbsis.categoria.Categoria;
-import br.com.hbsis.categoria.CategoriaDTO;
 import br.com.hbsis.categoria.CategoriaService;
-import br.com.hbsis.categoria.ICategoriaRepository;
-import br.com.hbsis.fornecedor.FornecedorService;
-import com.opencsv.*;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
-import javax.servlet.http.HttpServletResponse;
-import java.io.InputStreamReader;
-import java.io.PrintWriter;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -25,17 +12,12 @@ import java.util.Optional;
 public class LinhaCategoriaService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(LinhaCategoriaService.class);
-
     private final ILinhaCategoriaRepository iLinhaCategoriaRepository;
     private final CategoriaService categoriaService;
-    private final FornecedorService fornecedorService;
-    private final ICategoriaRepository iCategoriaRepository;
 
-    public LinhaCategoriaService(ILinhaCategoriaRepository iLinhaCategoriaRepository, CategoriaService categoriaService, FornecedorService fornecedorService, ICategoriaRepository iCategoriaRepository) {
+    public LinhaCategoriaService(ILinhaCategoriaRepository iLinhaCategoriaRepository, CategoriaService categoriaService) {
         this.iLinhaCategoriaRepository = iLinhaCategoriaRepository;
         this.categoriaService = categoriaService;
-        this.fornecedorService = fornecedorService;
-        this.iCategoriaRepository = iCategoriaRepository;
     }
 
     public LinhaCategoriaDTO save(LinhaCategoriaDTO linhaCategoriaDTO) {
@@ -60,16 +42,13 @@ public class LinhaCategoriaService {
         linhaCategoria = this.iLinhaCategoriaRepository.save(linhaCategoria);
         return LinhaCategoriaDTO.of(linhaCategoria);
     }
-
     public String validarCodigo(String codigo) {
         String codigoCorrigido = StringUtils.leftPad(codigo, 10, "0");
         return codigoCorrigido;
     }
-
     public List<LinhaCategoria> saveAll(List<LinhaCategoria> linhaCategorias) throws Exception {
         return iLinhaCategoriaRepository.saveAll(linhaCategorias);
     }
-
     public LinhaCategoriaDTO findByid(Long id) {
         Optional<LinhaCategoria> linhaCategoriaOptional = this.iLinhaCategoriaRepository.findById(id);
 
@@ -78,7 +57,6 @@ public class LinhaCategoriaService {
         }
         throw new IllegalArgumentException(String.format("ID %s não existe", id));
     }
-
     public LinhaCategoria findByLinhaCategoriaId(Long id) {
         Optional<LinhaCategoria> linhaCategoriaOptional = this.iLinhaCategoriaRepository.findById(id);
 
@@ -88,7 +66,6 @@ public class LinhaCategoriaService {
 
         throw new IllegalArgumentException(String.format("ID %s não existe", id));
     }
-
     public LinhaCategoria findByLinhaCategoriaCodLinhaCategoria(String codLinhaCategoria) {
         Optional<LinhaCategoria> optionalLinhaCategoria = this.iLinhaCategoriaRepository.findByCodLinhaCategoria(codLinhaCategoria);
 
@@ -98,7 +75,6 @@ public class LinhaCategoriaService {
 
         throw new IllegalArgumentException(String.format("codigo  %s don't exist", codLinhaCategoria));
     }
-
     public LinhaCategoriaDTO update(LinhaCategoriaDTO linhaCategoriaDTO, Long id) {
         Optional<LinhaCategoria> linhaCategoriaExistenteOptional = this.iLinhaCategoriaRepository.findById(id);
             this.validate(linhaCategoriaDTO);
@@ -125,89 +101,10 @@ public class LinhaCategoriaService {
         }
         throw new IllegalArgumentException(String.format("ID %s não existe", id));
     }
-
     public void delete(Long id) {
         LOGGER.info("Executando delete para linha de categoria de ID:[{}]", id);
         this.iLinhaCategoriaRepository.deleteById(id);
     }
-
-    //Faz a exportação
-    public void csvToLinhaCategoriaExport(HttpServletResponse response) throws Exception {
-        String nomeArquivo = "linhaCategorias.csv";
-        response.setContentType("text/csv");
-        response.setHeader(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + nomeArquivo + "\"");
-        PrintWriter writer = response.getWriter();
-
-        ICSVWriter icsvWriter = new CSVWriterBuilder(writer).withSeparator(';')
-                .withEscapeChar(CSVWriter.DEFAULT_ESCAPE_CHARACTER).withLineEnd(CSVWriter.DEFAULT_LINE_END).build();
-
-        String headerCSV[] = {"codigo", "nome", "codigo_categoria", "nome_categoria"};
-        icsvWriter.writeNext(headerCSV);
-
-        for (LinhaCategoria linha : iLinhaCategoriaRepository.findAll()) {
-            icsvWriter.writeNext(new String[]{linha.getCodLinhaCategoria(), linha.getNomeLinha(), linha.getCategoria().getCodigoCategoria(), linha.getCategoria().getNomeCategoria()}
-            );
-        }
-
-
-    }
-
-    //faz a importação
-    public List<LinhaCategoria> csvToLinhaCategoria(MultipartFile multipartFile) throws Exception {
-        InputStreamReader inputStreamReader = new InputStreamReader(multipartFile.getInputStream());
-
-        CSVReader csvReader = new CSVReaderBuilder(inputStreamReader)
-                .withSkipLines(1).build();
-
-
-        List<LinhaCategoria> leitura = new ArrayList<>();
-        List<String[]> linhas = csvReader.readAll();
-
-        for (String[] lista : linhas) {
-            try {
-
-                String[] resultado = lista[0].replaceAll("\"", "").split(";");
-
-                LinhaCategoria linhaCategoria = new LinhaCategoria();
-                Categoria categoria = new Categoria();
-
-                String codLinhacat = resultado[0];
-                String nomeLinha = resultado[1];
-                String codCategoria = resultado[2];
-
-                if (iLinhaCategoriaRepository.existsByCodLinhaCategoria(codLinhacat)) {
-                    LOGGER.info("Linha categoria: {}", codLinhacat + " já existe");
-                }
-
-                else if (iCategoriaRepository.existsByCodigoCategoria(codCategoria)) {
-
-                    linhaCategoria.setCodLinhaCategoria(codLinhacat);
-                    linhaCategoria.setNomeLinha(nomeLinha);
-                    categoria = categoriaService.findByCodigoCategoria(codCategoria);
-
-                    linhaCategoria.setCategoria(categoria);
-                    leitura.add(linhaCategoria);
-                }
-
-                 else if (!iLinhaCategoriaRepository.existsByCodLinhaCategoria(codLinhacat)){
-                    LOGGER.info("Linha categoria: {}", codLinhacat + " salvando");
-                    linhaCategoria.setCategoria(categoria);
-                    leitura.add(linhaCategoria);
-
-
-                }
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-
-        }
-        return iLinhaCategoriaRepository.saveAll(leitura);
-    }
-    public void importFromCsvlinhaCategoria(MultipartFile file) throws Exception {
-        List<LinhaCategoria> linhaCategorias = this.csvToLinhaCategoria(file);
-        this.saveAll(linhaCategorias);
-    }
-
     private void validate(LinhaCategoriaDTO linhaCategoriaDTO) {
         LOGGER.info("Validando linha da categoria");
         if (linhaCategoriaDTO == null) {
