+++ src/main/java/br/com/hbsis/produto/ProdutoRest.java
@@ -1,6 +1,7 @@
 package br.com.hbsis.produto;
 
 
+import br.com.hbsis.csv.CsvProduto;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,10 +15,13 @@ import javax.servlet.http.HttpServletResponse;
 public class ProdutoRest {
     private static final Logger LOGGER = LoggerFactory.getLogger(ProdutoRest.class);
     private final ProdutoService produtoService;
+    private final CsvProduto csvProduto;
 
     @Autowired
-    public ProdutoRest(ProdutoService produtoService) {
+    public ProdutoRest(ProdutoService produtoService, CsvProduto csvProduto) {
         this.produtoService = produtoService;
+        this.csvProduto = csvProduto;
+
     }
 
     @PostMapping
@@ -53,22 +57,22 @@ public class ProdutoRest {
 
     @GetMapping("/exporta-csv-produtos")
     public void findAll(HttpServletResponse response) throws Exception {
-        produtoService.csvToProdutoExport(response);
+        csvProduto.csvToProdutoExport(response);
     }
 
     @GetMapping("/exporta-csvproduto-fornecedor")
     public void findAllFornecedor(HttpServletResponse response) throws Exception {
-        produtoService.csvToProdutoFornecedorExport(response);
+        csvProduto.csvToProdutoFornecedorExport(response);
     }
 
     @PostMapping("/importa-csv-produtos")
     public void importCSV(@RequestParam("file") MultipartFile file) throws Exception {
-        produtoService.importFromCsvProduto(file);
+        csvProduto.importFromCsvProduto(file);
     }
 
     @PostMapping("/importa-csv-produtos-fornecedor/{id}")
     public void importCSVFornecedor(@PathVariable("id") Long id, @RequestParam("file") MultipartFile file) throws Exception {
-        produtoService.csvToProdutoFornecedor(id, file);
+        csvProduto.csvToProdutoFornecedor(id, file);
     }
 
 }
